web/requirements.txt
# Web app dependencies
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9   # PostgreSQL driver (ignored if using SQLite)

# Auth
bcrypt==4.1.3
python-jose[cryptography]==3.3.0

# Encryption
cryptography==42.0.7

# AI
anthropic>=0.28.0

# Email (IMAP + SMTP)
imapclient==2.3.1
beautifulsoup4==4.12.3

# iCal
icalendar==5.0.13

# HTTP
requests==2.32.3

# Env
python-dotenv==1.0.1
